Get fig_spec.rb to emit std(out|err) on failure.
Modified fig() to throw an exception by default.

Common scenario during development is that the fig external process will
fail for whatever reason, but the RSpec expectation is checking whether
a file was created, etc. meaning that we don't see stdout, stderr, etc.
but RSpec's failure message for the expectation, which isn't
informative.  Throwing an exception that RSpec will catch will correctly
integrate the fig output with the rest of the RSpec output.