Make os.download return true and add a test.
"os.download" was not returning true which was breaking repository.update_package. When using the "file" protocol, repository.install_package will not be called. Tests were passing because a publish will leave a package in the FIG_HOME as well as the FIG_REMOTE_URL. To test, comment out the "return true" in os.download and rerun the tests. You should see one test in fig_spec.rb break.