Performed minor refactor and improved error code.
Added rubymine files to .gitignore.
Changed instances of input to package_config_file.
Added rescue of RepositoryError in fig.rb so incorrect user assumptions will not generate a stack trace.
Provided a detailed comment explaining the ignoring of the FTPPermError in os.rb.
Changed the Logging of the FTP/Socket errors to only log the message since these are not errors within fig.
Changed the construction of the package object to allow passing of nil but performs a check which results in defaults if nil is passed.